Create Find the merge point of two liked lists
Harry is in the first stage of the triwizard tournament. He flies on his Firebolt broomstick while being followed by the Hungarian Horntail dragon. His path while flying can be represented using a linked list (Anything is possible with magic!), and so can be the path of the dragon. Harry wants to know how safe he is, i.e. he wants to know if there is a possibility that his path and the dragons path intersect. 

For example, if linked list A denotes harry's path and linked list B denotes the dragons path, the following situation would be dangerous for harry :
A:          a1 → a2
                   ↘
                     c1 → c2 → c3 → NULL
                   ↗
B:     b1 → b2 → b3
while the following situation is safe :

A:          a1 → a2 → a3 → NULL
                   
                   
B:     b1 → b2 → b3 → b4 → b5 → NULL